Moved last block inserted selector test from editor to the block-editor
